requirements.txt
 phidata
 streamlit 
 groq